refactor(component): Migrate ticket comment to use patternfly design system
ref: #170 #146 #160
